requirements.txt
requests==2.31.0
notion-client==2.2.1
woocommerce==3.0.0
python-dotenv==1.0.0
schedule==1.2.0
pydantic==2.5.0
loguru==0.7.2
